Pass dispatcher into the initialize function
Is there a reason for that in the TodoStore (todo example) , the "dispatcher" is not passed into the initialize function? 
So far i can see applies that only for the TodoStore. The rest of the the stores in this repository are passing the "dispatcher" into the initialize function.

The documentation does also not makes it any clearer.
https://github.com/yahoo/dispatchr#createstore